Move build and list subcommands as part of language-pack command
#155
